fix bug query order service

See merge request hai95.lv/Unicoin-Clothing-Website!57

+++ order-service/_infra/4_export_order_detail.sql
@@ -3,7 +3,7 @@ CREATE TABLE `export_order_detail` (
                                        `export_order_id` int DEFAULT NULL,
                                        `variant_id` int DEFAULT NULL,
                                        `quantity` int DEFAULT NULL,
-                                       `price` double DEFAULT NULL,
+                                       `price` int DEFAULT NULL,
                                        PRIMARY KEY (`export_order_detail_id`),
                                        KEY `fk_export_orders_idx` (`export_order_id`),
                                        CONSTRAINT `fk_export_orders` FOREIGN KEY (`export_order_id`) REFERENCES `export_orders` (`export_order_id`)
